Dockerfile: Remove VOLUME instruction
This instruction is usually useless e.g. when using novolume-plugin. In cloud environments volumes must be mounted and there's no point in keeping Prometheus data temporary volume dir only during life-time of a container. Another downside of using VOLUME instruction is it hard to undone it in child images and one would need to re-write the Dockerfile.