web/ui: disable React-based UI
The link to the new React-based UI was already removed from the OCP
version of Prometheus to avoid getting support requests for a feature
that is still experimental upstream. However it was possible to access
this UI by accessing the /graph/new URL manually.

This change removes completely the embedded React assets from the
generated binary. Someone trying to access /graph/new URL will
encounter a 404 error.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>