commit web ui assets to avoid yarn dependency
NOTE(spasquie): unlike upstream, we want the assets to be committed in
the repository because build environments don't have access to Yarn and
external package repositories.  This means that every time we update to
a new Prometheus version, 'make assets' should be run and the resulting
file committed to the repository.